+++ sims/verilator/testbench/include/cpu_tool.h
@@ -47,7 +47,7 @@ public:
     static int rf_trace_no_repeat;
     static int comp_pc_trace;
     static int comp_rf_trace;
-    static int dump_delay;
+    static int64_t dump_delay;
     static int dump_waveform;
     const static char* ram_file;
     const static char* data_vlog_file;
